Move workspace org setting to own file
