+++ packages/salesforcedx-vscode-core/src/context/index.ts
@@ -1,0 +1,4 @@
+export {
+  setupWorkspaceOrgType,
+  registerDefaultOrgWatcher
+} from './workspaceOrgType';
