+++ packages/salesforcedx-vscode-core/src/index.ts
@@ -7,11 +7,6 @@
 import * as path from 'path';
 import * as vscode from 'vscode';
 import { ConfigurationTarget } from 'vscode';
-
-import {
-  ForceConfigGet,
-  ForceOrgList
-} from '@salesforce/salesforcedx-utils-vscode/out/src/cli';
 import { channelService } from './channels';
 import {
   CompositeParametersGatherer,
@@ -68,6 +63,7 @@ import {
   SFDX_CLIENT_ENV_VAR,
   TERMINAL_INTEGRATED_ENVS
 } from './constants';
+import { registerDefaultOrgWatcher, setupWorkspaceOrgType } from './context';
 import * as decorators from './decorators';
 import { nls } from './messages';
 import { isDemoMode } from './modes/demo-mode';
@@ -341,66 +337,14 @@ function registerCommands(
   );
 }
 
-async function setupWorkspaceOrgType() {
-  const defaultUsername = await getDefaultUsername();
-  const defaultUsernameIsSet = typeof defaultUsername !== 'undefined';
-
-  let isScratchOrg = false;
-  if (defaultUsernameIsSet) {
-    const forceOrgList = new ForceOrgList();
-    isScratchOrg = await forceOrgList.isScratchOrg(defaultUsername!);
-  }
-  vscode.commands.executeCommand(
-    'setContext',
-    'sfdx:default_org_is_scratch_org',
-    defaultUsernameIsSet && isScratchOrg
+function registerIsvAuthWatcher(): vscode.Disposable {
+  const isvAuthWatcher = vscode.workspace.createFileSystemWatcher(
+    path.join('.sfdx', 'sfdx-config.json')
   );
-  vscode.commands.executeCommand(
-    'setContext',
-    'sfdx:default_org_is_non_scratch_org',
-    defaultUsernameIsSet && !isScratchOrg
-  );
-}
-
-async function getDefaultUsername(): Promise<string | undefined> {
-  if (
-    vscode.workspace.workspaceFolders instanceof Array &&
-    vscode.workspace.workspaceFolders.length > 0
-  ) {
-    const forceConfig = await new ForceConfigGet().getConfig(
-      vscode.workspace.workspaceFolders[0].uri.fsPath,
-      'defaultusername'
-    );
-    return forceConfig.get('defaultusername');
-  }
-}
-
-function registerSfdxConfigWatcher(
-  callbacks: Array<() => void>,
-  context: vscode.ExtensionContext
-): vscode.Disposable | undefined {
-  if (
-    vscode.workspace.workspaceFolders instanceof Array &&
-    vscode.workspace.workspaceFolders.length > 0
-  ) {
-    const sfdxConfigWatcher = vscode.workspace.createFileSystemWatcher(
-      new vscode.RelativePattern(
-        vscode.workspace.workspaceFolders[0],
-        path.join('.sfdx', 'sfdx-config.json')
-      )
-    );
-    sfdxConfigWatcher.onDidChange(uri => {
-      callbacks.forEach(fn => fn());
-    });
-    sfdxConfigWatcher.onDidCreate(uri => {
-      callbacks.forEach(fn => fn());
-    });
-    sfdxConfigWatcher.onDidDelete(uri => {
-      callbacks.forEach(fn => fn());
-    });
-    context.subscriptions.push(sfdxConfigWatcher);
-  }
-  return;
+  isvAuthWatcher.onDidChange(uri => setupGlobalDefaultUserIsvAuth());
+  isvAuthWatcher.onDidCreate(uri => setupGlobalDefaultUserIsvAuth());
+  isvAuthWatcher.onDidDelete(uri => setupGlobalDefaultUserIsvAuth());
+  return vscode.Disposable.from(isvAuthWatcher);
 }
 
 export async function activate(context: vscode.ExtensionContext) {
@@ -448,8 +392,6 @@ export async function activate(context: vscode.ExtensionContext) {
     sfdxProjectOpened
   );
 
-  // Set context variables based on .sfdx/sfdx-config.json
-  const contextSettingFunctions = [];
   const sfdxApexDebuggerExtension = vscode.extensions.getExtension(
     'salesforce.salesforcedx-vscode-apex-debugger'
   );
@@ -463,12 +405,13 @@ export async function activate(context: vscode.ExtensionContext) {
     sfdxApexDebuggerExtension &&
     sfdxApexDebuggerExtension.id
   ) {
-    // Setup default user for CLI this is done in core because it shares access
-    // to GlobalCliEnvironment with the commands
+    console.log('Setting up ISV Debugger environment variables');
+    // register watcher for ISV authentication and setup default user for CLI
+    // this is done in core because it shares access to GlobalCliEnvironment with the commands
     // (VS Code does not seem to allow sharing npm modules between extensions)
     try {
-      console.log('Setting up ISV Debugger environment variables');
-      contextSettingFunctions.push(setupGlobalDefaultUserIsvAuth);
+      context.subscriptions.push(registerIsvAuthWatcher());
+      console.log('Configured file watcher for .sfdx/sfdx-config.json');
       await setupGlobalDefaultUserIsvAuth();
     } catch (e) {
       console.error(e);
@@ -478,12 +421,9 @@ export async function activate(context: vscode.ExtensionContext) {
     }
   }
 
-  // Setup default workspace org type
-  contextSettingFunctions.push(setupWorkspaceOrgType);
+  // Set context for default workspace org
   await setupWorkspaceOrgType();
-
-  registerSfdxConfigWatcher(contextSettingFunctions, context);
-  console.log('Configured file watcher for .sfdx/sfdx-config.json');
+  registerDefaultOrgWatcher(context);
 
   // Commands
   const commands = registerCommands(context);
